+++ infrastructure/src/resources/cloud/index.ts
@@ -1,9 +1,11 @@
-import { Domain, Project } from '@pulumi/digitalocean'
+import { Domain, Project, KubernetesCluster, DropletSlug } from '@pulumi/digitalocean'
+import { DigitalOceanCluster } from './DigitalOceanCluster'
 
 export default class {
     public provisionAll(): void {
         const domain: Domain = this.provisionDomain()
-        this.provisionProject(domain)
+        const cluster: KubernetesCluster = this.provisionCluster()
+        this.provisionProject(domain, cluster)
     }
 
     private provisionDomain(): Domain {
@@ -12,14 +14,32 @@ export default class {
         })
     }
 
-    private provisionProject(domain: Domain): Project {
+    private provisionCluster(): KubernetesCluster {
+        const cluster = new DigitalOceanCluster()
+
+        return new KubernetesCluster('main-cluster', {
+            name: cluster.name(),
+            version: cluster.version(),
+            region: 'fra1',
+            autoUpgrade: false,
+            nodePool: {
+                name: 'worker',
+                size: DropletSlug.DropletS1VCPU1GB,
+                autoScale: false,
+                nodeCount: 1
+            }
+        })
+    }
+
+    private provisionProject(domain: Domain, cluster: KubernetesCluster): Project {
         return new Project('main-project', {
             name: 'VFM',
             environment: 'Production',
             description: 'Vehicle Fleet Management infrastructure',
             purpose: 'Web Application',
             resources: [
-                domain.domainUrn
+                domain.domainUrn,
+                cluster.clusterUrn
             ]
         })
     }